ActiveRecord::TestCase: reap all leaked connection on teardown
Some tests do leak connections, which cause some other tests to stall
for a very long time when they call `clear_all_connections`.

Each leaked connection takes 5 seconds to be cleared because of the
checkout timeout.

So on teardown to inspect all pools to make sure nothing was leaked
and if some connections were leaked we cleanup the state.