Allow raw CTEs in with query method
Support for the `.with` query method was added in https://github.com/rails/rails/pull/37944

The hash based syntax added in that PR supports Relations, Arel
SelectManagers, and SqlLiterals as expressions, but only allows simple names
for the CTEs.

In some circumstances, it's useful to be able to name the columns of the
CTE as well as the table.

For example, if the expression is a ValuesList:

    WITH magic(a, b, c) AS (VALUES (2, 6, 7), (9, 5, 1), (4, 3, 8))
    SELECT a + b + c FROM magic;

... we need to be able to name the columns in order to reference them
later.

One use case for having this supported in ActiveRecord's with query
method is joining the results from a previous query to a subsequent one.
For example, if you were walking through an org chart, but you couldn't
use recursive CTEs, you could use this approach to keep information
about the management path:

    big_bosses = Employee.where(manager: nil).pluck(id)

    first_level = Employee.with(Arel::Nodes::Cte.new(
        Arel.sql("bosses(id)"),
        Arel::Nodes::Grouping.new(Arel::Nodes::ValuesList.new(big_bosses))
      )
      .joins("INNER JOIN employees ON emplooyees.manager = bosses.id")
      .pluck("employees.id", "bosses.id")

    second_level = Employee.with(Arel::Nodes::Cte.new(
        Arel.sql("bosses(id, grand_boss_id)"),
        Arel::Nodes::Grouping.new(Arel::Nodes::ValuesList.new(first_level))
      )
      .joins("INNER JOIN employees ON emplooyees.manager = bosses.id")
      .pluck("employees.id", "bosses.id", "bosses.grand_boss_id")

    ... and so on ... (not tested)

I'm not sure about this use case particularly, but I imagine there are a
few other situations where it's useful to have more control over the
name of the CTE than we currently have.