Support IN queries on array columns
This commit fixes a bug (or at least an inconsistency) that made it
possible to submit IN queries on array columns.

A non-array column can be queried via `where(name: [value1, value2])`
which is automatically converted to `name IN (value1, value2)` in SQL.
An array column would treat that array as the value to search for, so
it'd be natural to expect an array of arrays would result in an `IN`
query.
Unfortunately, that wasn't the case as `where(name: [[value1], [value2]])`
would return `name = {{value1}, {value2}}` in SQL, instead of `name IN ({value1}, {value2})`.

That behavior stemmed from the fact that Active Record would force
equality comparison for array columns if the value was an array.
In order to allow `IN` queries this check was change to check whether
it's a **one-dimensional** array by checking that the first item of is not an array.

After these changes the following logic applies for querying array
columns:

1. Using a one-dimensional array results in an ordinary equality query.
2. Using a two-dimensional array produces an `IN` query.