[Fix #51242] Rework in_batches(use_ranges: true) to be more efficient
When `use_ranges: true` option is used, we do not need to return the
whole list of primary keys for each range. This wastes network resources
and takes longer than necessary, as only the last primary key from the
range would be needed.

We can instead use OFFSET to peek at the last primary key of the next
batch, instead of returning the whole list of pks from the range.

There is a trade-off, however, as we need an additional query to confirm
the size of the very last batch, and the value of the last last primary
key. Unless we only have a handful of small batches, this strategy is a
winner overall, as it reduces the time and network resources spent
generating batches.