Extract `ActionText::TrixEditor`
Related to [rails/actiontext#41](https://github.com/rails/actiontext/issues/41)

The introduction of the `ActionText::TrixEditor` class enables the
deprecation of a variety of class- and instance-level methods across the
`ActionText` namespace. Action Text benefits from deprecating methods in
order to reduce its public API.

There are two categories of deprecations in this changeset:
`module`/`class` deprecations and method deprecations.

The `module` and `class` deprecations target modules that are internal
and "private" in intent, but public in Ruby and not marked as `:nodoc:`.
They're either particular to Trix, or their responsibilities have been
given to `ActionText::Editor`.

The method deprecations also target methods that are internal and
"private" in intent, but `public` in Ruby and not marked as `:nodoc:`.
In general, determining whether or not to deprecate a method hinged on
whether its name or arguments mentioned `"trix"`, *or* if it lacked
method-level documentation or comments of any kind.

When possible, the previous code paths have been deprecated *and*
re-implemented in terms of an `ActionText::TrixEditor` instance.

Additions
---

The main aim of this changeset is to provide a single, extensible
entrypoint for third-party editors (that are not Trix) to integrate with
Action Text.

The initial responsibilities of the `ActionText::TrixEditor` have been
determined by consolidating a variety of (class- and instance-) methods
across a variety of classes to reduce the number of touchpoints for
applications and engines that provide third-party editor integrations.
The vast majority of its method definitions are direct copy-and-pastes
from their original sources. Any method that previously mentioned `trix`
in its name or argument list has replaced and generalized that
occurrence with `editor`.

Classes that inherit from `ActionText::Editor` can override two methods
that accept and return `Fragment` instances:

* `Editor#as_canonical` method accepts an editor-sourced
  `ActionText::Fragment` instance, transforms it, then returns an
  `ActionText::Fragment` instance to be stored.

* `Editor#as_editable` method accepts a storage-sourced
  `ActionText::Fragment` instance, transforms it, then returns an
  `ActionText::Fragment` instance to be edited.

The interface utilizes `Fragment` instances rather than `Content` or
String instances.

The `Fragment` class serves as an agnostic layer on top of HTML and
Plain Text string manipulation through Nokogiri. It has a simple
interface of transforms:

* `find_all(selector)`
* `update(&block)`
* `replace(selector, &block)`

Accepting **and** returning a transformation-ready `Fragment` enables
a more consistent adapter interface.

Changes
---

*These changes aim to be considered implementation details, and intend
to remain "private" from an API perspective.* Future commits can be made
to expand upon and document the responsibilities of the adapter's
interface.