railties: configure sanitizer vendor in 7.1 defaults more robustly
Note: this is a manual backport of #51267 as it did not apply cleanly.

rails-html-santizer is a dependency of Action View and a transitive
dependency of Action Text (via Action Pack), but may not be loaded
until after railties sets configuration defaults.

This change `require`s rails-html-sanitizer immediately before it's
needed, and avoids the possibly-incorrect assumption that
Rails::HTML::Sanitizer is already defined.

Closes rails#51246

Co-authored-by: Rafael Mendonça França <rafael@rubyonrails.org>
Co-authored-by: Mike Dalessio <mike.dalessio@gmail.com>