Do not build View watcher until the first `updated?` check
Currently initialization of every Rails::Engine leads to the creation
of a new View watcher when the engine prepends its paths.
This contributes to the time it takes to perform the first cold request
on a lazy loaded application.

This change delays the initialization of the View watcher until the first
`updated?` check is performed.

Co-Authored-By: Gannon McGibbon <gannon.mcgibbon@gmail.com>