Expire caching when a download fail while proxying in ActiveStorage Fix #51284
Both Proxy controllers in Active Storage set the caching headers early
before streaming.

In some instances (see #51284), it is possible for the file
download (from the service to server) to fail before we send the first
byte to the client (response not yet committed).

In those instances, this change would invalidate the cache and return
a better response status before closing the stream.