Use dartsass, generate files to the output dir
`dartsass-rails` is part of the Rails Gemfile, and thus should always
be available. Use `bundle exec` to make sure that's the version we're
using, always.

Generate the files directly to the output dir, so they don't go to the
source assets folder as intermediate files, otherwise we'd need to
explicitly tell git to ignore them. We don't need the intermediate files
stored there anyway, so let's just generate directly to the final output
folder, which is already git ignored.

Also remove `stylesrc` folder/files from the output folder, as it's not
necessary there.

Finally, remove the style and highlight css files, as they're going to
be generated from scss now.