Set the selected Rails version based on the generated guides version
Instead of setting it dynamically via JS, use the known version that the
guide is being generated with, or edge, to determine the selected Rails
version in the version dropdown.

Since each version is set when generating the guides, we can "hardcode"
the selected version upon guides generation, instead of using JS to set
it dynamically.