Improve contrast on dark mode
The red shades used on links, heading and other elements didn't have enough contrast with
the dark background.

To maintain the brand identity, I kept the red shades but made the background color darker.
I tried not to change the color too much, so I used the lightest shade of gray that would get
at least 3:1 on a contrast ratio checker.

Fixes #51363