Reduce contrast of text and inline code elements.
To much contrast in dark mode can cause halation and makes
text hard to read.

The Accessible Perceptual Contrast Algorithm (APCA)[1] has
preliminary guidelines ([2]) for maximum contrast in dark mode.

The changes introduced in this PR ensure that the contrast for
text content does not exceed Lc 90 and does not exceed Lc 75 for
large text.

The background color for inline code elements was also darkened
so that there is less contrast with the surrounding text.

1. https://github.com/Myndex
2. https://github.com/Myndex/SAPC-APCA/discussions/106