Add label and accessible content
Having a select element without a label outside a form isn't
directly problematic, but the text Version is the label of the
select and should act as such.

Because the select element is set up to browse to a different page
which loses the current focus position, users must be instructed
about this context change. It would be much better to display this
content to everyone, but having this for screenreaders only is
better than not having it at all.