Fix accessible text for voice commands
Using the title attribute means some screenreaders will not be able
to show the expanded accessible text at all, but using aria-label
loses the hover text currently available.

If the title attribute is still used, the accessible label should
start with the same content as the visible label. This is necessary
for most voice operated tooling which only match starting text.

Having "Return" as the start of the accessible label and "Guides"
as the start of the visible label means in most cases users that
rely on voice commands will not be able to browse here.