Add landmarks to all guide pages
Landmarks such as header, section, footer, article, and the main
content landmark help navigating when using assistive tooling.

Since each guide is a self-contained article, a header section
with the guide introduction and navigation, a footer section which
is not part of the content but a call to action, and the rest can
be clearly separated, landmarks here are appropiate.