Consistent scrolling behaviour respecting pref
Use the preference to detect if a user wants reduced motion and if
not enable smooth scrolling. This takes care of the back-to-top
smooth scrolling, and also adds it to anchor links. If this is
unwanted, the change can be reverted and a matchmedia check should
be done in JavaScript instead before smooth scrolling