Add skip links to quickly navigate past navs
Skip links are a common practice to allow keyboard and other users
to skip past large blocks of content that are present on every
page, which in this case are the secondary navigation at the top,
the version selector, and the chapter navigation.

The skip links only show when focused.