ActiveRecord suppport authentication with non-postgres database
It's possible to configure PostgreSQL user in a way that doesn't
allow it to connect using "postgres" database but instead only
his own database. For more info see `pg_hba.conf`

Currently such users wouldn't be able to be used with ActiveRecord
so fix that by allowing this.

Note that permission to connect and actual permission to query "postgres"
are separate and independent things.