Add an internal API to trigger association loading asynchronously
This is not yet a public API, I'd like to experiment with a few
ideas and run it by a few people first. But that's pretty much all
the necessary plumbing.

One big limitation right now is with `has_many through`, as it
requires two queries back to back. I'm not sure how to best
handle that, but I think worst case we could at least trigger
the first of the two, that would already be a win.