Update builder doc links to point to the rails repo [ci skip]
After reviewing the Action View Overview guide, a discussion started
around the builder gem that was detailed in the guide. We decided to
move it to the rails/ umbrella so we can help maintain it as needed,
since the framework uses/documents it, keeping it similar to jbuilder.

https://github.com/rails/rails/pull/51435#discussion_r1556390070

cherry-pick of d462fb54b459e222efc2e2480397af44a9c55361 from main.