Don't mess with Minitest unless RAILS_ENV is set
Minitest will automatically scan all installed gems and load plugins
from those gems.  We should detect whether or not we're being run within
the context of a Rails app and only change MT behavior in that case.

To determine if we're being run via `bin/rails` we'll just check if
RAILS_ENV is set.

Ref: https://github.com/minitest/minitest/issues/996
Ref: https://github.com/minitest/minitest/issues/725