Fix select helper with block returning non-string
As suggested by the docs for [FormOptionsHelper#select][1], the helper
can be called with a block to customize option tag attributes:

    select(report, :campaign_ids) do
      available_campaigns.each do |c|
        tag.option(c.name, value: c.id, data: { tags: c.tags.to_json })
      end
    end

Prior to this commit, a NoMethodError error would be raised when
available_campaigns in the above example was the empty array. This
happened because the block passed to select would return a non-string
value (the empty array) causing the call to capture in
Helpers::Tags::Select#initialize to return nil:

    def initialize(object_name, method_name, template_object, choices, options, html_options)
      @choices = block_given? ? template_object.capture { yield || "" } : choices

This commit fixes this by moving the `|| ""` outside the block passed to
capture.

[1]: https://api.rubyonrails.org/classes/ActionView/Helpers/FormOptionsHelper.html#method-i-select