Fix vendor_trix not fetching css file
The issue is the difference between how Trix packages the 1.x and 2.x
series. In 1.x, there is only a trix.js file but in 2.x there is
trix.esm.js and trix.umd.js.

Since Rails 7 is staying on Trix 1.x, the vendor_trix script needed to
be updated since there is no esm.js or umd.js files in Trix 1.x